poster_ai_api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asn1crypto=0.24.0=py37_0
blas=1.0=mkl
ca-certificates=2019.1.23=0
certifi=2018.11.29=py37_0
cffi=1.11.5=py37h74b6da3_1
chardet=3.0.4=py37_1
cryptography=2.5=py37h7a1dbc1_0
cycler=0.10.0=py37_0
dataclasses=0.6=pypi_0
django=2.1.7=py37_0
django-filter=2.1.0=pypi_0
djangorestframework=3.9.1=pypi_0
djongo=1.2.31=pypi_0
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py37_0
intel-openmp=2019.1=144
jpeg=9b=hb83a4c4_2
kiwisolver=1.0.1=py37h6538335_0
libpng=1.6.36=h2a8f88b_0
libtiff=4.0.10=hb898794_2
matplotlib=3.0.2=py37hc8f65d3_0
mkl=2019.1=144
mkl_fft=1.0.10=py37h14836fe_0
mkl_random=1.0.2=py37h343c172_0
numpy=1.15.4=py37h19fb1c0_0
numpy-base=1.15.4=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1a=he774522_0
pillow=5.4.1=py37hdc69c19_0
pip=19.0.1=py37_0
pycparser=2.19=py37_0
pymongo=3.7.2=pypi_0
pyopenssl=19.0.0=py37_0
pyparsing=2.3.1=py37_0
pyqt=5.9.2=py37h6538335_2
pysocks=1.6.8=py37_0
python=3.7.2=h8c8aaf0_2
python-dateutil=2.7.5=py37_0
pytz=2018.9=py37_0
qt=5.9.7=vc14h73c81de_0
requests=2.21.0=py37_0
setuptools=40.7.3=py37_0
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
sqlite=3.26.0=he774522_0
sqlparse=0.2.4=pypi_0
tk=8.6.8=hfa6e2cd_0
tornado=5.1.1=py37hfa6e2cd_0
urllib3=1.24.1=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_0
wheel=0.32.3=py37_0
win_inet_pton=1.0.1=py37_1
wincertstore=0.2=py37_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
